pyproject.toml
[tool.poetry]
name = "spotter"
version = "0.1.0"
description = ""
authors = ["pablobuenaposada <pablo.buenaposada@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-elasticsearch-dsl = "^8.0"
django-elasticsearch-dsl-drf = "^0.22.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.6"
black = "^24.8.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
model-bakery = "^1.19.5"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
lint.select = [
    "F", # Pyflakes
    "E", # pycodestyle
    "W", # Warning
    "C90", # mccabe
    "I",  # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "Q", # flake8-quotes
    "SIM", # flake8-simplify
    "PGH", # pygrep-hooks
    "PLE", # Error
    "PLW", # Warning
    "RSE", # flake8-raise
]
lint.ignore = ["E501"]